chore(theme): relocate token build helpers to scripts/
Move refs, resolve and css-vars helpers from tokens/build/ to a
top-level scripts/ folder so they sit alongside upcoming compile
pipelines and are decoupled from the token data tree.
