+++ packages/theme/src/tokens/index.js
@@ -2,7 +2,7 @@
  * Tokens public exports.
  */
 
-export { createCssVars, cssVarsString, injectCssVars } from './build/css-vars.js';
-export * from './primitives/colors.js';
+export { createCssVars, cssVarsString, injectCssVars } from '../scripts/css-vars.js';
+export * from './primitives/colors/colors.js';
 export * from './semantic/colors.js';
 export * from './theme.js';
